+++ ts/src/lib/actions/unknown.ts
@@ -1,3 +1,4 @@
+import { $Bytes } from "kryo/builtins/bytes";
 import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { IntegerType } from "kryo/types/integer";
@@ -9,12 +10,14 @@ import { $ActionType, ActionType } from "../action-type";
 export interface Unknown extends ActionBase {
   action: ActionType.Unknown;
   code: Uint8;
+  data: Uint8Array;
 }
 
 export const $Unknown: DocumentIoType<Unknown> = new DocumentType<Unknown>({
   properties: {
     action: {type: new LiteralType({type: $ActionType, value: ActionType.Unknown as ActionType.Unknown})},
     code: {type: new IntegerType()},
+    data: {type: $Bytes},
   },
   changeCase: CaseStyle.SnakeCase,
 });