+++ ts/src/test/raw.spec.ts
@@ -12,10 +12,12 @@ const JSON_READER: JsonReader = new JsonReader();
 
 const sampleNames: ReadonlyArray<string> = [
   "define-function",
-  "trace",
   "end",
   "error",
   "error-with-message",
+  "trace",
+  "unknown",
+  "unknown-with-data",
 ];
 
 describe("$Action.read", function () {